pyproject.toml
[tool.poetry]
name = "pytorch-transforms-illustrated"
version = "0.1.0"
description = ""
authors = ["diegulio <diego.machado@usach.cl>"]
readme = "README.md"
packages = [{include = "pytorch_transforms_illustrated"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.27.2"
torchvision = "^0.16.0"
pillow = "^10.1.0"
black = "^23.10.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
